+++ packages/tolk/lib/uop/tolk_uop.mli
@@ -71,6 +71,7 @@ module Axis_type = Axis_type
 module Dtype = Dtype
 module Const = Const
 module Bound = Bound
+module Target = Target
 module Uop = Uop
 module Render = Render
 module Upat = Upat